OBPIH-7827 Refactor pack list item import functionality: rename package, add PackImportItemCommand, and update controller and service references
