+++ grails-app/controllers/org/pih/warehouse/picklist/PicklistController.groovy
@@ -214,41 +214,6 @@ class PicklistController {
         }
     }
 
-    def exportPackTemplate() {
-        String format = params.get("format", "csv")
-
-        List<Map<String, String>> lineItems = stockMovementService.buildPackTemplateLineItems(params.id)
-
-        String fileName = "PackListItems\$-${params.id}-template"
-
-        switch (format) {
-            case "csv":
-                String csv = dataService.generateCsv(lineItems)
-                response.setHeader("Content-disposition", "attachment; filename=\"${fileName}.csv\"")
-                render(contentType: "text/csv", text: csv, encoding: "UTF-8")
-                break
-            case "xls":
-                response.contentType = "application/vnd.ms-excel"
-                response.setHeader("Content-disposition", "attachment; filename=\"${fileName}.xls\"")
-                documentService.generateExcel(response.outputStream, lineItems)
-                response.outputStream.flush()
-                break
-            default:
-                throw new IllegalFormatException("Unable to determine the proper rendering format for request for format ${format}")
-        }
-    }
-
-    def importPackListItems(PackImportDataCommand command) {
-        command.location = command.location ?: Location.get(session.warehouse.id)
-        List<String> errors = stockMovementService.processPackListImport(command, params.id)
-
-        if (!errors.isEmpty()) {
-            render([message: "Data imported with errors", errors: errors] as JSON)
-        }
-
-        render([message: "Data imported successfully"] as JSON)
-    }
-
     def importPickListItems(PicklistImportDataCommand command) {
 
         // Bind stock movement based on provided ID
